289. Game of Life
